+++ index.js
@@ -4,7 +4,8 @@ import {AppRegistry} from 'react-native';
 import App from './App';
 import {name as appName} from './app.json';
 import Storybook from './storybook';
+import {LOAD_STORYBOOK} from './env.json';
 
-const Component = __DEV__ ? Storybook : App;
+const Component = LOAD_STORYBOOK ? Storybook : App;
 
 AppRegistry.registerComponent(appName, () => Component);