+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2023-12-06
 ### Added
 
 - Added progress bar component
 - Added base button component
+- Added list component
+- Added possibility to change between app mode and storybook app mode
 
 ## [1.1.1] - 2023-10-31